+++ lib/src/models/chat_message.dart
@@ -1,10 +1,11 @@
-part of dash_chat_2;
+part of '../../dash_chat_2.dart';
 
 /// {@category Models}
 class ChatMessage {
   ChatMessage({
     required this.user,
     required this.createdAt,
+    this.isMarkdown = false,
     this.text = '',
     this.medias,
     this.quickReplies,
@@ -20,6 +21,7 @@ class ChatMessage {
       user: ChatUser.fromJson(jsonData['user'] as Map<String, dynamic>),
       createdAt: DateTime.parse(jsonData['createdAt'].toString()).toLocal(),
       text: jsonData['text']?.toString() ?? '',
+      isMarkdown: jsonData['isMarkdown']?.toString() == 'true',
       medias: jsonData['medias'] != null
           ? (jsonData['medias'] as List<dynamic>)
               .map((dynamic media) =>
@@ -46,6 +48,9 @@ class ChatMessage {
     );
   }
 
+  /// If the message is Markdown formatted then it will be converted to Markdown (by default it will be false)
+  bool isMarkdown;
+
   /// Text of the message (optional because you can also just send a media)
   String text;
 
@@ -66,7 +71,7 @@ class ChatMessage {
   /// Date of the message
   DateTime createdAt;
 
-  /// Mentionned elements in the message
+  /// Mentioned elements in the message
   List<Mention>? mentions;
 
   /// Status of the message TODO:
@@ -89,6 +94,7 @@ class ChatMessage {
       'mentions': mentions,
       'status': status.toString(),
       'replyTo': replyTo?.toJson(),
+      'isMarkdown': isMarkdown,
     };
   }
 }
@@ -104,6 +110,10 @@ class MessageStatus {
     switch (value) {
       case 'none':
         return MessageStatus.none;
+      case 'failed':
+        return MessageStatus.failed;
+      case 'sent':
+        return MessageStatus.sent;
       case 'read':
         return MessageStatus.read;
       case 'received':
@@ -116,6 +126,8 @@ class MessageStatus {
   }
 
   static const MessageStatus none = MessageStatus._internal('none');
+  static const MessageStatus failed = MessageStatus._internal('failed');
+  static const MessageStatus sent = MessageStatus._internal('sent');
   static const MessageStatus read = MessageStatus._internal('read');
   static const MessageStatus received = MessageStatus._internal('received');
   static const MessageStatus pending = MessageStatus._internal('pending');