+++ lib/src/widgets/input_toolbar/default_send_button.dart
@@ -1,18 +1,19 @@
-part of dash_chat_2;
+part of '../../../dash_chat_2.dart';
 
 /// {@category Default widgets}
 Widget Function(Function send) defaultSendButton({
   required Color color,
   IconData icon = Icons.send,
   EdgeInsets? padding,
+  bool disabled = false,
 }) =>
     (Function fct) => InkWell(
-          onTap: () => fct(),
+          onTap: disabled ? null : () => fct(),
           child: Padding(
             padding: padding ??
                 const EdgeInsets.symmetric(horizontal: 8, vertical: 10),
             child: Icon(
-              Icons.send,
+              icon,
               color: color,
             ),
           ),