+++ lib/src/widgets/message_row/default_message_text.dart
@@ -1,4 +1,4 @@
-part of dash_chat_2;
+part of '../../../dash_chat_2.dart';
 
 /// {@category Default widgets}
 class DefaultMessageText extends StatelessWidget {
@@ -66,13 +66,20 @@ class DefaultMessageText extends StatelessWidget {
         match
             .groups(List<int>.generate(match.groupCount, (int i) => i + 1))
             .forEach((String? part) {
+          Mention? mention;
           if (mentionRegex.hasMatch(part!)) {
-            Mention mention = message.mentions!.firstWhere(
-              (Mention m) => m.title == part,
-            );
+            try {
+              mention = message.mentions?.firstWhere(
+                (Mention m) => m.title == part,
+              );
+            } catch (e) {
+              // There is no mention
+            }
+          }
+          if (mention != null) {
             res.add(getMention(context, mention));
           } else {
-            res.add(getParsePattern(context, part));
+            res.add(getParsePattern(context, part, message.isMarkdown));
           }
         });
         if (res.isNotEmpty) {
@@ -80,22 +87,35 @@ class DefaultMessageText extends StatelessWidget {
         }
       }
     }
-    return <Widget>[getParsePattern(context, message.text)];
+    return <Widget>[getParsePattern(context, message.text, message.isMarkdown)];
   }
 
-  Widget getParsePattern(BuildContext context, String text) {
-    return ParsedText(
-      parse: messageOptions.parsePatterns != null
-          ? messageOptions.parsePatterns!
-          : defaultPersePatterns,
-      text: text,
-      alignment: isOwnMessage ? TextAlign.end : TextAlign.start,
-      style: TextStyle(
-        color: isOwnMessage
-            ? messageOptions.currentUserTextColor(context)
-            : messageOptions.textColor,
-      ),
-    );
+  Widget getParsePattern(BuildContext context, String text, bool isMarkdown) {
+    return isMarkdown
+        ? MarkdownBody(
+            data: text,
+            selectable: true,
+            alignment: isOwnMessage ? TextAlign.end : TextAlign.start,
+            onTapLink: (String value, String? href, String title) {
+              if (href != null) {
+                openLink(href);
+              } else {
+                openLink(value);
+              }
+            },
+          )
+        : ParsedText(
+            parse: messageOptions.parsePatterns != null
+                ? messageOptions.parsePatterns!
+                : defaultParsePatterns,
+            text: text,
+            alignment: isOwnMessage ? TextAlign.end : TextAlign.start,
+            style: TextStyle(
+              color: isOwnMessage
+                  ? messageOptions.currentUserTextColor(context)
+                  : messageOptions.textColor,
+            ),
+          );
   }
 
   Widget getMention(BuildContext context, Mention mention) {